Extract digitalocean resources to a module
This is the first pass at trying providers as a module and I am not
sure I like this approach.

I think a better approach will be to have folders for each provider so
you can more or less choose the deployment target, then push out any
common resources into a shared module folder that you interact with.